+++ src/image/index.android.ts
@@ -1,7 +1,6 @@
 export * from './index-common';
 import { Background, Color, ImageAsset, ImageSource, Trace, Utils, backgroundInternalProperty, knownFolders, path } from '@nativescript/core';
 import { isString } from '@nativescript/core/utils/types';
-import { layout } from '@nativescript/core/utils/layout-helper';
 import {
     AnimatedImage,
     CLog,
@@ -50,27 +49,29 @@ const globalSignatureKey = 'v1';
 export function initialize(config?: ImagePipelineConfigSetting): void {
     if (!initialized) {
         const context = Utils.android.getApplicationContext();
-        if (!context) {
-            return;
+        const glideConfig = com.nativescript.image.GlideConfiguration.getInstance();
+        if (config?.memoryCacheSize > 0) {
+            glideConfig.setMemoryCacheSize(config.memoryCacheSize);
         }
-        if (config?.usePersistentCacheKeyStore) {
-            const sharedStore = new com.nativescript.image.SharedPrefCacheKeyStore(context.getApplicationContext());
-            com.nativescript.image.EvictionManager.get().setPersistentStore(sharedStore);
-        }
-        // bumping `v1` will invalidate all cache
-        signature = new com.bumptech.glide.signature.ObjectKey(config?.globalSignatureKey ?? globalSignatureKey);
 
-        initialized = true;
+        if (config?.memoryCacheScreens > 0) {
+            glideConfig.setMemoryCacheScreens(config.memoryCacheScreens);
+        }
+
+        if (config?.usePersistentCacheKeyStore) {
+            com.nativescript.image.EvictionManager.get().setPersistentStore(new com.nativescript.image.SharedPrefCacheKeyStore(Utils.android.getApplicationContext()));
+        }
+
+        signature = new com.bumptech.glide.signature.ObjectKey(config?.globalSignatureKey || globalSignatureKey);
+
+        // Now initialize Glide (which will read from GlideConfiguration)
         glideInstance = com.bumptech.glide.Glide.get(context);
-        com.nativescript.image.EvictionManager.get().clearAll();
 
         // this is needed for further buildKey to trigger ...
-        com.bumptech.glide.Glide.with(context)
-            .load('toto?ts=' + Date().valueOf())
-            .apply(new com.bumptech.glide.request.RequestOptions().signature(new com.bumptech.glide.signature.ObjectKey(Date().valueOf())))
-            .preload();
+        com.bumptech.glide.Glide.with(context).load('toto').apply(new com.bumptech.glide.request.RequestOptions().signature(signature)).preload();
         // com.nativescript.image.EngineKeyFactoryMethodDumper.dumpKeyFactoryMethods(glideInstance);
         // com.nativescript.image.ForcePreloadTest.forcePreloadAfterInjection(context, 'https://example.com/test-image.png');
+        initialized = true;
     }
 }
 
@@ -113,13 +114,13 @@ export class ImagePipeline {
                 url,
                 new com.nativescript.image.EvictionManager.EvictionCallback({
                     onComplete(success: boolean, error) {
-                        if (success) {
-                            resolve();
-                        } else {
+                        if (error) {
                             if (Trace.isEnabled()) {
                                 CLog(CLogTypes.error, error);
                             }
                             reject(error);
+                        } else {
+                            resolve();
                         }
                     }
                 })
@@ -134,13 +135,13 @@ export class ImagePipeline {
                 url,
                 new com.nativescript.image.EvictionManager.EvictionCallback({
                     onComplete(success: boolean, error) {
-                        if (success) {
-                            resolve();
-                        } else {
+                        if (error) {
                             if (Trace.isEnabled()) {
                                 CLog(CLogTypes.error, error);
                             }
                             reject(error);
+                        } else {
+                            resolve();
                         }
                     }
                 })
@@ -155,13 +156,13 @@ export class ImagePipeline {
                 url,
                 new com.nativescript.image.EvictionManager.EvictionCallback({
                     onComplete(success: boolean, error) {
-                        if (success) {
-                            resolve();
-                        } else {
+                        if (error) {
                             if (Trace.isEnabled()) {
                                 CLog(CLogTypes.error, error);
                             }
                             reject(error);
+                        } else {
+                            resolve();
                         }
                     }
                 })
@@ -174,13 +175,13 @@ export class ImagePipeline {
             com.nativescript.image.EvictionManager.get().clearAll(
                 new com.nativescript.image.EvictionManager.EvictionCallback({
                     onComplete(success: boolean, error) {
-                        if (success) {
-                            resolve();
-                        } else {
+                        if (error) {
                             if (Trace.isEnabled()) {
                                 CLog(CLogTypes.error, error);
                             }
                             reject(error);
+                        } else {
+                            resolve();
                         }
                     }
                 })
@@ -231,40 +232,7 @@ export class ImagePipeline {
     }
 
     prefetchToMemoryCache(uri: string): Promise<void> {
-        return new Promise((resolve, reject) => {
-            try {
-                const context = Utils.android.getApplicationContext();
-                const requestManager = com.bumptech.glide.Glide.with(context);
-                const requestBuilder = requestManager.asBitmap().load(uri);
-
-                let futureTarget = null;
-                function clearLater() {
-                    if (futureTarget) {
-                        setTimeout(() => {
-                            com.bumptech.glide.Glide.with(context).clear(futureTarget);
-                        }, 0);
-                    }
-                }
-
-                const listener = new com.bumptech.glide.request.RequestListener({
-                    onLoadFailed(e: any, model: any, target: any, isFirstResource: boolean): boolean {
-                        clearLater();
-                        reject(e);
-                        return true; // consumed
-                    },
-                    onResourceReady(resource: any, model: any, target: any, dataSource: any, isFirstResource: boolean): boolean {
-                        clearLater();
-                        resolve();
-                        return true; // consumed
-                    }
-                });
-
-                // Kick the request off and keep a reference to the FutureTarget so it can be cleared.
-                futureTarget = requestBuilder.listener(listener).preload();
-            } catch (error) {
-                reject(error);
-            }
-        });
+        return this.prefetchToCache(uri, false);
     }
 
     private prefetchToCache(uri: string, toDiskCache: boolean): Promise<void> {
@@ -272,7 +240,15 @@ export class ImagePipeline {
             try {
                 const context = Utils.android.getApplicationContext();
                 const requestManager = com.bumptech.glide.Glide.with(context);
-                const requestBuilder = toDiskCache ? requestManager.downloadOnly().load(uri) : requestManager.asBitmap().load(uri);
+                const loadModel = new com.nativescript.image.CustomGlideUrl(
+                    uri,
+                    null,
+                    null, // Can be null
+                    null // Can be null
+                );
+                const requestBuilder = (toDiskCache ? requestManager.downloadOnly().load(uri) : requestManager.asBitmap().load(loadModel)).apply(
+                    new com.bumptech.glide.request.RequestOptions().signature(signature)
+                );
 
                 let futureTarget = null;
                 function clearLater() {
@@ -620,23 +596,23 @@ export class Img extends ImageBase {
             });
         }
         // Use CustomGlideUrl if we need headers, progress, or load source
+        let headersMap;
         if (this.isNetworkRequest && (this.headers || this.progressCallback || this.loadSourceCallback)) {
-            const headersMap = new java.util.HashMap();
+            headersMap = new java.util.HashMap();
 
             if (this.headers) {
                 for (const key in this.headers) {
                     headersMap.put(key, this.headers[key]);
                 }
             }
-
-            loadModel = new com.nativescript.image.CustomGlideUrl(
-                uri,
-                headersMap,
-                this.progressCallback, // Can be null
-                this.loadSourceCallback // Can be null
-            );
         }
-        requestBuilder = com.bumptech.glide.Glide.with(context).load(loadModel).signature(new com.bumptech.glide.signature.ObjectKey(Date().valueOf()));
+        loadModel = new com.nativescript.image.CustomGlideUrl(
+            uri,
+            headersMap,
+            this.progressCallback, // Can be null
+            this.loadSourceCallback // Can be null
+        );
+        requestBuilder = com.bumptech.glide.Glide.with(context).load(loadModel);
 
         // Apply transformations (blur, rounded corners, etc.)
         const transformations = [];
@@ -720,14 +696,13 @@ export class Img extends ImageBase {
         }
 
         const owner = new WeakRef(this);
-        const sourceKey = new com.bumptech.glide.signature.ObjectKey(uri);
         const objectArr = Array.create(com.bumptech.glide.request.RequestListener, 2);
         const ro = new com.bumptech.glide.request.RequestOptions().signature(signature);
 
         objectArr[0] = new com.nativescript.image.SaveKeysRequestListener(
             uri,
-            uri,
-            sourceKey,
+            loadModel,
+            new com.bumptech.glide.signature.ObjectKey(uri), // fallback only
             signature,
             this.decodeWidth || com.bumptech.glide.request.target.Target.SIZE_ORIGINAL,
             this.decodeHeight || com.bumptech.glide.request.target.Target.SIZE_ORIGINAL,
@@ -817,7 +792,7 @@ export class Img extends ImageBase {
             target.setClearFirst(false);
         }
 
-        requestBuilder.signature(signature).listener(new com.nativescript.image.CompositeRequestListener(objectArr)).into(target);
+        requestBuilder.apply(ro).listener(new com.nativescript.image.CompositeRequestListener(objectArr)).into(target);
     }
 
     private notifyLoadSource(source: string) {